move QueryContext to query_context module, some doc improvements
